requirements.txt
anyio==4.6.2.post1
asgiref==3.7.2
beautifulsoup4==4.11.1
blinker==1.9.0
bs4==0.0.1
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.6
Django==4.2.7
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-filter==23.5
django-jasmine==0.5.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.8.0
Flask==3.1.0
fonttools==4.47.0
googletrans==3.0.0
gTTS==2.4.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==1.0.6
httpx==0.27.2
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==3.0.2
matplotlib==3.8.2
numpy==1.26.2
opencv-python==4.11.0.86
packaging==23.2
pillow==10.2.0
platformdirs==2.5.4
psycopg2-binary==2.9.6
PyJWT==2.10.1
pyparsing==3.1.1
pyTelegramBotAPI==4.24.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==21.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SpeechRecognition==3.10.1
sqlparse==0.4.4
telebot==0.0.5
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.16.7
Werkzeug==3.1.3
